improved move_library function

+++ rust/src/api/library.rs
@@ -440,13 +440,19 @@ pub fn init_library(app_data_dir: String) {
 
 #[flutter_rust_bridge::frb]
 pub async fn move_library(new_dir: String) -> Result<()> {
-    use std::env;
+    use std::{env, fs};
 
 
     let old_path = PathBuf::from(env::var("APP_DATA_DIR")?).join("songbook");
     let new_path = PathBuf::from(&new_dir).join("songbook");
-    copy_all(&old_path, new_path)?;
-    remove_all(old_path)?;
+    if let Err(_) = fs::rename(&old_path, &new_path) {
+        if let Err(err) = copy_all(&old_path, &new_path) {
+            remove_all(new_path)?;
+            return Err(err);
+        } else {
+            remove_all(old_path)?;
+        }
+    }
 
 
     env::set_var("APP_DATA_DIR", new_dir);
@@ -455,7 +461,7 @@ pub async fn move_library(new_dir: String) -> Result<()> {
 fn remove_all<P: AsRef<Path>>(path: P) -> std::io::Result<()> {
     if path.as_ref().is_file() {
         std::fs::remove_file(path)?;
-    } else {
+    } else if path.as_ref().is_dir() {
         std::fs::remove_dir_all(path)?;
     }
